Subclass check for isa compatibility.
